requirements.txt
torch == 1.10.1
torchvision == 0.11.2
numpy == 1.24.3
tqdm == 4.65.0
timm == 0.6.12
beautifulsoup4 == 4.11.2
einops == 0.6.0
h5py == 3.7.0
jupyter == 1.0.0
matplotlib == 3.7.1
openpyxl == 3.0.10
Pillow == 9.4.0
opencv-python == 4.7.0.72

